+++ Foosball/Foosball.csproj
@@ -57,26 +57,27 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="WriteReadData.cs" />
-    <Compile Include="Counting Goals.cs">
+    <Compile Include="Utils\Team.cs" />
+    <Compile Include="Utils\WriteReadData.cs" />
+    <Compile Include="forms\Counting Goals.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Counting Goals.Designer.cs">
+    <Compile Include="forms\Counting Goals.Designer.cs">
       <DependentUpon>Counting Goals.cs</DependentUpon>
     </Compile>
-    <Compile Include="Leaderboards.cs">
+    <Compile Include="forms\Leaderboards.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Leaderboards.Designer.cs">
+    <Compile Include="forms\Leaderboards.Designer.cs">
       <DependentUpon>Leaderboards.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="timer1.cs" />
-    <Compile Include="ScoreInput.cs">
+    <Compile Include="forms\ScoreInput.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ScoreInput.Designer.cs">
+    <Compile Include="forms\ScoreInput.Designer.cs">
       <DependentUpon>ScoreInput.cs</DependentUpon>
     </Compile>
     <Compile Include="Start Screen.cs">
@@ -85,10 +86,10 @@
     <Compile Include="Start Screen.Designer.cs">
       <DependentUpon>Start Screen.cs</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="Counting Goals.resx">
+    <EmbeddedResource Include="forms\Counting Goals.resx">
       <DependentUpon>Counting Goals.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Leaderboards.resx">
+    <EmbeddedResource Include="forms\Leaderboards.resx">
       <DependentUpon>Leaderboards.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -104,7 +105,7 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
-    <EmbeddedResource Include="ScoreInput.resx">
+    <EmbeddedResource Include="forms\ScoreInput.resx">
       <DependentUpon>ScoreInput.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Start Screen.resx">
@@ -129,6 +130,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\Plus.jpg" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\video_files\VID_20171020_100012.mp4" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Emgu.CV.3.2.0.2721\build\Emgu.CV.targets" Condition="Exists('..\packages\Emgu.CV.3.2.0.2721\build\Emgu.CV.targets')" />
