+++ Foosball/Properties/Resources.resx
@@ -124,4 +124,7 @@
   <data name="Plus" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\Plus.jpg;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
+  <data name="VID_20171020_100012" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\video_files\VID_20171020_100012.mp4;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
 </root>
