+++ Foosball/Utils/Team.cs
@@ -1,0 +1,15 @@
+﻿namespace Foosball.Utils
+{
+    public class Team
+    {
+        public string TeamName { get; set; }
+        public int GlobalScore { get; set; }
+
+        public Team(string teamName, int globalScore)
+        {
+            TeamName = teamName;
+            GlobalScore = globalScore;
+        }
+
+    }
+}
