+++ Foosball/forms/Counting Goals.resx
