+++ Foosball/forms/ScoreInput.resx
